Pulled pi_creature constructs out into their own folder, so as to note some distinction from the remainder of the library
